Let the three- and four-pile splitters share one board client
Their BoardClients were the same ~100 lines twice over. The JSX differed
only in the pile index list, which `range(board.length)` covers, and in
the border rule: three piles drew a divider between piles 0 and 1 on
whichever was taller, four piles did the same for both pairs, so
`pileId % 2 === 0 && board[pileId] >= board[pileId + 1]` and its mirror
reproduce both — the odd last pile of a three-pile board compares against
a missing neighbour and gets no divider, as before.

The selection now lives in `useMoveScopedState`, which is what mid-turn
UI state is for: it drops the half-made choice on any move, so the reset
inside the deferred split goes away, and a selection no longer survives
an undo or a restart. It also means the board is inert for the beat
between the player's own two moves, where it used to leave some pieces
clickable.

`pile-splitter` keeps its own client — two piles make the discarded one
implicit, so its turn is a single click — but takes the same `pieceColor`
helper. It was stacking `bg-blue-800` and its overrides on one element,
where which colour wins is Tailwind's emit order rather than the order
written; and it showed a bare `0` for the pile the bot had just emptied,
where the siblings show the bin.
